change code of conduct link
Changes the Checklist for Featured Products to say the alternative to making your own code of conduct is linking to the Open Austin code of conduct, instead of linking to the Open Austin Values document. The values statement isn't really a substitute for a code of conduct, and the values document has some random unorganized notes at the bottom that most people don't need to read.